Fix pointer truncation in fopScnM_CreateReq i_data

+++ include/f_op/f_op_scene_mng.h
@@ -10,7 +10,7 @@ typedef struct base_process_class base_process_class;
 scene_class* fopScnM_SearchByID(fpc_ProcID id);
 int fopScnM_ChangeReq(scene_class* i_scene, s16 i_procName, s16 param_3, u16 param_4);
 fpc_ProcID fopScnM_DeleteReq(scene_class* i_scene);
-int fopScnM_CreateReq(s16 i_procName, s16 param_2, u16 param_3, u32 i_data);
+int fopScnM_CreateReq(s16 i_procName, s16 param_2, u16 param_3, uintptr_t i_data);
 u32 fopScnM_ReRequest(s16 i_procName, u32 i_data);
 void fopScnM_Management();
 void fopScnM_Init();
